+++ tests/conftest.py
@@ -64,6 +64,15 @@ def graph_with_nodes() -> HeteroData:
 
 
 @pytest.fixture
+def graph_with_isolated_nodes() -> HeteroData:
+    graph = HeteroData()
+    graph["test_nodes"].x = torch.tensor([[1], [2], [3], [4], [5], [6]])
+    graph["test_nodes"]["mask_attr"] = torch.tensor([[1], [1], [1], [0], [0], [0]], dtype=torch.bool)
+    graph["test_nodes", "to", "test_nodes"].edge_index = torch.tensor([[2, 3, 4], [1, 2, 3]])
+    return graph
+
+
+@pytest.fixture
 def graph_nodes_and_edges() -> HeteroData:
     """Graph with 1 set of nodes and edges."""
     coords = np.array([[lat, lon] for lat in lats for lon in lons])
@@ -91,10 +100,9 @@ def config_file(tmp_path) -> tuple[str, str]:
             {
                 "source_name": "test_nodes",
                 "target_name": "test_nodes",
-                "edge_builder": {
-                    "_target_": "anemoi.graphs.edges.KNNEdges",
-                    "num_nearest_neighbours": 3,
-                },
+                "edge_builders": [
+                    {"_target_": "anemoi.graphs.edges.KNNEdges", "num_nearest_neighbours": 3},
+                ],
                 "attributes": {
                     "dist_norm": {"_target_": "anemoi.graphs.edges.attributes.EdgeLength"},
                     "edge_dirs": {"_target_": "anemoi.graphs.edges.attributes.EdgeDirection"},